refactor: move OcppTagRepository into AuthTagService
